requirements.txt
pandas>=2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
oracledb
pymysql==1.1.0
pyinstaller>=6.1.0
